Moving data related files to data module
